pyproject.toml
[project]
name = "cursor-conversation-rag"
version = "0.1.0"
description = "Portable conversation-history RAG system optimized for multilingual retrieval"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=2.3.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "tiktoken>=0.5.2",
    "python-dotenv>=1.0.0",
    "mcp>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/conversation_rag"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
